Added Waspeteers (community) section to landing page.
