+++ web/src/pages/index.jsx
@@ -2,6 +2,7 @@ import Head from "@docusaurus/Head";
 import useDocusaurusContext from "@docusaurus/useDocusaurusContext";
 
 import AI from "../components/AI";
+import ExampleApps from "../components/ExampleApps";
 import Faq from "../components/Faq";
 import FeaturesExplorer from "../components/FeaturesExplorer";
 import Footer from "../components/Footer";
@@ -14,7 +15,9 @@ import OpenSaas from "../components/OpenSaas";
 import Philosophy from "../components/Philosophy";
 import Properties from "../components/Properties";
 import Roadmap from "../components/Roadmap";
-import WaspOutThere from "../components/WaspOutThere";
+import ShippedWithWasp from "../components/ShippedWithWasp";
+import Testimonials from "../components/Testimonials";
+import Waspeteers from "../components/Waspeteers";
 
 import waspCoverPhoto from "../../static/img/wasp_twitter_cover.png";
 import "./index.css";
@@ -56,7 +59,10 @@ const Index = () => {
             <AI />
             <HowItWorks />
             <NoLockIn />
-            <WaspOutThere />
+            <ExampleApps />
+            <ShippedWithWasp />
+            <Testimonials />
+            <Waspeteers />
             <Philosophy />
             <OpenSaas />
             <Newsletter />
